+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to the extension are documented in this file.
 
 ## [Unreleased]
 
+- Adding verbose debug output during su file reading to understand individual problems better.
+
 ## [0.3.0] - 2022-08-11
 
 - Support GCC<10 where .su file entries do not contain full paths.